requirements.txt
asgiref==3.8.1
bleach==6.1.0
Django==4.2.13
django-jazzmin==3.0.0
django-simple-history==3.7.0
psycopg2==2.9.9
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
webencodings==0.5.1
